Refactored whole Project structure, removed .gitattributes changed pom.xml to not create a dep-reduced-pom and changed src directory to default maven src directory
