+++ src/main/java/com/projectkorra/rpg/ProjectKorraRPG.java
@@ -3,8 +3,8 @@ package com.projectkorra.rpg;
 import com.projectkorra.rpg.commands.HelpCommand;
 import com.projectkorra.rpg.commands.RPGCommandBase;
 import com.projectkorra.rpg.configuration.ConfigManager;
-import com.projectkorra.rpg.modules.ModuleManager;
 import com.projectkorra.rpg.metrics.MetricsLite;
+import com.projectkorra.rpg.modules.ModuleManager;
 import net.luckperms.api.LuckPerms;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.RegisteredServiceProvider;
