+++ src/main/java/com/projectkorra/rpg/RPGMethods.java
