+++ src/main/java/com/projectkorra/rpg/modules/elementassignments/manager/AssignmentManager.java
