+++ src/main/java/com/projectkorra/rpg/modules/elementassignments/util/AssignmentGroup.java
