+++ src/main/java/com/projectkorra/rpg/modules/randomavatar/listeners/AvatarListener.java
