+++ src/main/java/com/projectkorra/rpg/modules/randomavatar/manager/AvatarManager.java
