+++ src/main/java/com/projectkorra/rpg/modules/worldevents/listeners/WorldEventScheduleListener.java
