+++ src/main/java/com/projectkorra/rpg/modules/worldevents/util/display/bossbar/WorldEventBossBar.java
