+++ src/main/java/com/projectkorra/rpg/modules/worldevents/util/display/scoreboard/ScoreboardDisplay.java
