+++ src/main/java/com/projectkorra/rpg/modules/worldevents/util/schedule/WorldEventScheduleStrategy.java
