+++ src/main/java/com/projectkorra/rpg/modules/worldevents/util/schedule/WorldEventScheduleStrategyFactory.java
@@ -1,5 +1,6 @@
 package com.projectkorra.rpg.modules.worldevents.util.schedule;
 
+import com.projectkorra.rpg.modules.worldevents.util.schedule.WorldEventScheduleStrategy;
 import com.projectkorra.rpg.modules.worldevents.util.schedule.strategies.AtTimeStrategy;
 import com.projectkorra.rpg.modules.worldevents.util.schedule.strategies.EveryInGameDaysStrategy;
 import com.projectkorra.rpg.modules.worldevents.util.schedule.strategies.EveryRealWorldDaysStrategy;
