+++ src/main/java/com/projectkorra/rpg/modules/worldevents/util/schedule/WorldEventScheduler.java
@@ -3,6 +3,8 @@ package com.projectkorra.rpg.modules.worldevents.util.schedule;
 import com.projectkorra.rpg.ProjectKorraRPG;
 import com.projectkorra.rpg.modules.worldevents.WorldEvent;
 import com.projectkorra.rpg.modules.worldevents.listeners.WorldEventScheduleListener;
+import com.projectkorra.rpg.modules.worldevents.util.schedule.WorldEventScheduleStrategy;
+import com.projectkorra.rpg.modules.worldevents.util.schedule.WorldEventScheduleStrategyFactory;
 import org.bukkit.plugin.Plugin;
 
 import java.util.*;
